ollama/ats_checker/requirements.txt
requests>=2.31.0
PyPDF2>=3.0.1
python-docx>=0.8.11
docx2txt>=0.8
